// src/app/Models/Progress.ts
export class Progress{
    Course: String;
    Class: String;
    CreditHrs : number;
    Folder: String;
    Files: Number;
}

export class Dates{
    Quizes: string[];
    Assignments: string[];
    Lectures: string[];
    midsPaper: string;
    finalsPaper: string;
}

export class ChartData{
    Class: String;
    Percentage: Number[];
}